+++ src/test/java/com/nexmo/client/voice/ListCallsMethodTest.java
@@ -68,8 +68,7 @@ public class ListCallsMethodTest {
 
     @Test
     public void makeRequestWithFilter() throws Exception {
-        CallsFilter callsFilter = new CallsFilter();
-        callsFilter.setPageSize(3);
+        CallsFilter callsFilter = new CallsFilter.Builder().pageSize(3).build();
         RequestBuilder request = method.makeRequest(callsFilter);
         assertEquals("GET", request.getMethod());
         assertEquals("https://api.nexmo.com/v1/calls?page_size=3", request.getUri().toString());
@@ -132,8 +131,7 @@ public class ListCallsMethodTest {
                 .build()));
 
         try {
-            CallsFilter filter = new CallsFilter();
-            filter.setPageSize(30);
+            CallsFilter filter = new CallsFilter.Builder().pageSize(30).build();
             RequestBuilder request = method.makeRequest(filter);
             // Anything past here only executes if our assertion is incorrect:
             LOG.error("SnsRequest URI: " + request.getUri());
@@ -150,8 +148,7 @@ public class ListCallsMethodTest {
 
     @Test
     public void testDefaultUri() throws Exception {
-        CallsFilter filter = new CallsFilter();
-        filter.setPageSize(3);
+        CallsFilter filter = new CallsFilter.Builder().pageSize(3).build();
 
         RequestBuilder builder = method.makeRequest(filter);
         assertEquals("GET", builder.getMethod());
@@ -162,8 +159,7 @@ public class ListCallsMethodTest {
     public void testCustomUri() throws Exception {
         HttpWrapper wrapper = new HttpWrapper(new HttpConfig.Builder().baseUri("https://example.com").build());
         ListCallsMethod method = new ListCallsMethod(wrapper);
-        CallsFilter filter = new CallsFilter();
-        filter.setPageSize(3);
+        CallsFilter filter = new CallsFilter.Builder().pageSize(3).build();
 
         RequestBuilder builder = method.makeRequest(filter);
         assertEquals("GET", builder.getMethod());