+++ src/test/java/com/nexmo/client/voice/VoiceClientTest.java
@@ -57,10 +57,7 @@ public class VoiceClientTest {
 
         byte[] keyBytes = testUtils.loadKey("test/keys/application_key");
         AuthCollection authCollection = new AuthCollection();
-        authCollection.add(new JWTAuthMethod(
-                "951614e0-eec4-4087-a6b1-3f4c2f169cb0",
-                keyBytes
-        ));
+        authCollection.add(new JWTAuthMethod("951614e0-eec4-4087-a6b1-3f4c2f169cb0", keyBytes));
 
         HttpWrapper wrapper = new HttpWrapper(authCollection);
         wrapper.setHttpClient(client);
@@ -70,93 +67,59 @@ public class VoiceClientTest {
 
     @Test
     public void testCreateCall() throws Exception {
-        VoiceClient client = new VoiceClient(stubHttpWrapper(200, "{\n" +
-                "  \"conversation_uuid\": \"63f61863-4a51-4f6b-86e1-46edebio0391\",\n" +
-                "  \"status\": \"started\",\n" +
-                "  \"direction\": \"outbound\"\n" +
-                "}"));
-        CallEvent evt = client.createCall(
-                new Call("447700900903", "447700900904", "http://api.example.com/answer"));
+        VoiceClient client = new VoiceClient(stubHttpWrapper(200,
+                "{\n" + "  \"conversation_uuid\": \"63f61863-4a51-4f6b-86e1-46edebio0391\",\n"
+                        + "  \"status\": \"started\",\n" + "  \"direction\": \"outbound\"\n" + "}"
+        ));
+        CallEvent evt = client.createCall(new Call("447700900903", "447700900904", "http://api.example.com/answer"));
         assertEquals("63f61863-4a51-4f6b-86e1-46edebio0391", evt.getConversationUuid());
     }
 
     @Test
     public void testListCallsNoFilter() throws Exception {
-        VoiceClient client = new VoiceClient(stubHttpWrapper(200, "{\n" +
-                "  \"page_size\": 10,\n" +
-                "  \"record_index\": 0,\n" +
-                "  \"count\": 0,\n" +
-                "  \"_embedded\": {\n" +
-                "    \"calls\": []\n" +
-                "  },\n" +
-                "  \"_links\": {\n" +
-                "    \"self\": {\n" +
-                "      \"href\": \"/v1/calls?page_size=10&record_index=0\"\n" +
-                "    },\n" +
-                "    \"first\": {\n" +
-                "      \"href\": \"/v1/calls?page_size=10\"\n" +
-                "    },\n" +
-                "    \"last\": {\n" +
-                "      \"href\": \"/v1/calls?page_size=10\"\n" +
-                "    }\n" +
-                "  }\n" +
-                "}\n"));
+        VoiceClient client = new VoiceClient(stubHttpWrapper(200,
+                "{\n" + "  \"page_size\": 10,\n" + "  \"record_index\": 0,\n" + "  \"count\": 0,\n"
+                        + "  \"_embedded\": {\n" + "    \"calls\": []\n" + "  },\n" + "  \"_links\": {\n"
+                        + "    \"self\": {\n" + "      \"href\": \"/v1/calls?page_size=10&record_index=0\"\n"
+                        + "    },\n" + "    \"first\": {\n" + "      \"href\": \"/v1/calls?page_size=10\"\n"
+                        + "    },\n" + "    \"last\": {\n" + "      \"href\": \"/v1/calls?page_size=10\"\n" + "    }\n"
+                        + "  }\n" + "}\n"
+        ));
         CallInfoPage page = client.listCalls();
         assertEquals(0, page.getCount());
     }
 
     @Test
     public void testListCallsWithFilter() throws Exception {
-        VoiceClient client = new VoiceClient(stubHttpWrapper(200, "{\n" +
-                "  \"page_size\": 10,\n" +
-                "  \"record_index\": 0,\n" +
-                "  \"count\": 0,\n" +
-                "  \"_embedded\": {\n" +
-                "    \"calls\": []\n" +
-                "  },\n" +
-                "  \"_links\": {\n" +
-                "    \"self\": {\n" +
-                "      \"href\": \"/v1/calls?page_size=10&record_index=0\"\n" +
-                "    },\n" +
-                "    \"first\": {\n" +
-                "      \"href\": \"/v1/calls?page_size=10\"\n" +
-                "    },\n" +
-                "    \"last\": {\n" +
-                "      \"href\": \"/v1/calls?page_size=10\"\n" +
-                "    }\n" +
-                "  }\n" +
-                "}\n"));
-        CallInfoPage page = client.listCalls(new CallsFilter());
+        VoiceClient client = new VoiceClient(stubHttpWrapper(200,
+                "{\n" + "  \"page_size\": 10,\n" + "  \"record_index\": 0,\n" + "  \"count\": 0,\n"
+                        + "  \"_embedded\": {\n" + "    \"calls\": []\n" + "  },\n" + "  \"_links\": {\n"
+                        + "    \"self\": {\n" + "      \"href\": \"/v1/calls?page_size=10&record_index=0\"\n"
+                        + "    },\n" + "    \"first\": {\n" + "      \"href\": \"/v1/calls?page_size=10\"\n"
+                        + "    },\n" + "    \"last\": {\n" + "      \"href\": \"/v1/calls?page_size=10\"\n" + "    }\n"
+                        + "  }\n" + "}\n"
+        ));
+        CallInfoPage page = client.listCalls(new CallsFilter.Builder().build());
         assertEquals(0, page.getCount());
     }
 
     @Test
     public void testGetCallDetails() throws Exception {
-        VoiceClient client = new VoiceClient(stubHttpWrapper(200, "      {\n" +
-                "        \"uuid\": \"93137ee3-580e-45f7-a61a-e0b5716000ef\",\n" +
-                "        \"status\": \"completed\",\n" +
-                "        \"direction\": \"outbound\",\n" +
-                "        \"rate\": \"0.02400000\",\n" +
-                "        \"price\": \"0.00280000\",\n" +
-                "        \"duration\": \"7\",\n" +
-                "        \"network\": \"23410\",\n" +
-                "        \"conversation_uuid\": \"aa17bd11-c895-4225-840d-30dc38c31e50\",\n" +
-                "        \"start_time\": \"2017-01-13T13:55:02.000Z\",\n" +
-                "        \"end_time\": \"2017-01-13T13:55:09.000Z\",\n" +
-                "        \"to\": {\n" +
-                "          \"type\": \"phone\",\n" +
-                "          \"number\": \"447700900104\"\n" +
-                "        },\n" +
-                "        \"from\": {\n" +
-                "          \"type\": \"phone\",\n" +
-                "          \"number\": \"447700900105\"\n" +
-                "        },\n" +
-                "        \"_links\": {\n" +
-                "          \"self\": {\n" +
-                "            \"href\": \"/v1/calls/93137ee3-580e-45f7-a61a-e0b5716000ef\"\n" +
-                "          }\n" +
-                "        }\n" +
-                "      }\n"));
+        VoiceClient client = new VoiceClient(stubHttpWrapper(200,
+                "      {\n" + "        \"uuid\": \"93137ee3-580e-45f7-a61a-e0b5716000ef\",\n"
+                        + "        \"status\": \"completed\",\n" + "        \"direction\": \"outbound\",\n"
+                        + "        \"rate\": \"0.02400000\",\n" + "        \"price\": \"0.00280000\",\n"
+                        + "        \"duration\": \"7\",\n" + "        \"network\": \"23410\",\n"
+                        + "        \"conversation_uuid\": \"aa17bd11-c895-4225-840d-30dc38c31e50\",\n"
+                        + "        \"start_time\": \"2017-01-13T13:55:02.000Z\",\n"
+                        + "        \"end_time\": \"2017-01-13T13:55:09.000Z\",\n" + "        \"to\": {\n"
+                        + "          \"type\": \"phone\",\n" + "          \"number\": \"447700900104\"\n"
+                        + "        },\n" + "        \"from\": {\n" + "          \"type\": \"phone\",\n"
+                        + "          \"number\": \"447700900105\"\n" + "        },\n" + "        \"_links\": {\n"
+                        + "          \"self\": {\n"
+                        + "            \"href\": \"/v1/calls/93137ee3-580e-45f7-a61a-e0b5716000ef\"\n" + "          }\n"
+                        + "        }\n" + "      }\n"
+        ));
         CallInfo call = client.getCallDetails("93137ee3-580e-45f7-a61a-e0b5716000ef");
         assertEquals("93137ee3-580e-45f7-a61a-e0b5716000ef", call.getUuid());
     }
@@ -164,10 +127,9 @@ public class VoiceClientTest {
     @Test
     public void testSendDtmf() throws Exception {
         VoiceClient client = new VoiceClient(stubHttpWrapper(200,
-                "{\n" +
-                        "  \"message\": \"DTMF sent\",\n" +
-                        "  \"uuid\": \"ssf61863-4a51-ef6b-11e1-w6edebcf93bb\"\n" +
-                        "}"));
+                "{\n" + "  \"message\": \"DTMF sent\",\n" + "  \"uuid\": \"ssf61863-4a51-ef6b-11e1-w6edebcf93bb\"\n"
+                        + "}"
+        ));
 
         DtmfResponse response = client.sendDtmf("ssf61863-4a51-ef6b-11e1-w6edebcf93bb", "332393");
         assertEquals("ssf61863-4a51-ef6b-11e1-w6edebcf93bb", response.getUuid());
@@ -184,27 +146,31 @@ public class VoiceClientTest {
     @Test
     public void testModifyCall2() throws Exception {
         VoiceClient client = new VoiceClient(stubHttpWrapper(200, "{\"message\":\"Received\"}"));
-        ModifyCallResponse call = client.modifyCall(
-                new CallModifier("93137ee3-580e-45f7-a61a-e0b5716000ef", ModifyCallAction.MUTE));
+        ModifyCallResponse call = client.modifyCall(new CallModifier("93137ee3-580e-45f7-a61a-e0b5716000ef",
+                ModifyCallAction.MUTE
+        ));
         assertEquals("Received", call.getMessage());
     }
 
     @Test
     public void testTransferCall() throws Exception {
         VoiceClient client = new VoiceClient(stubHttpWrapper(200, "{\"message\":\"Received\"}"));
-        ModifyCallResponse call = client.transferCall(
-                "93137ee3-580e-45f7-a61a-e0b5716000ef", "https://example.com/ncco2");
+        ModifyCallResponse call = client.transferCall("93137ee3-580e-45f7-a61a-e0b5716000ef",
+                "https://example.com/ncco2"
+        );
         assertEquals("Received", call.getMessage());
     }
 
     @Test
     public void testStartStreamNonLooping() throws Exception {
         VoiceClient client = new VoiceClient(stubHttpWrapper(200,
-                "{\n" +
-                        "  \"message\": \"Stream started\",\n" +
-                        "  \"uuid\": \"ssf61863-4a51-ef6b-11e1-w6edebcf93bb\"\n" +
-                        "}"));
-        StreamResponse response = client.startStream("ssf61863-4a51-ef6b-11e1-w6edebcf93bb","https://nexmo-community.github.io/ncco-examples/assets/voice_api_audio_streaming.mp3");
+                "{\n" + "  \"message\": \"Stream started\",\n"
+                        + "  \"uuid\": \"ssf61863-4a51-ef6b-11e1-w6edebcf93bb\"\n" + "}"
+        ));
+        StreamResponse response = client.startStream(
+                "ssf61863-4a51-ef6b-11e1-w6edebcf93bb",
+                "https://nexmo-community.github.io/ncco-examples/assets/voice_api_audio_streaming.mp3"
+        );
         assertEquals("Stream started", response.getMessage());
         assertEquals("ssf61863-4a51-ef6b-11e1-w6edebcf93bb", response.getUuid());
     }
@@ -212,21 +178,24 @@ public class VoiceClientTest {
     @Test
     public void testStartStreamLooping() throws Exception {
         VoiceClient client = new VoiceClient(stubHttpWrapper(200,
-                "{\n" +
-                        "  \"message\": \"Stream started\",\n" +
-                        "  \"uuid\": \"ssf61863-4a51-ef6b-11e1-w6edebcf93bb\"\n" +
-                        "}"));
-        StreamResponse response = client.startStream("ssf61863-4a51-ef6b-11e1-w6edebcf93bb","https://nexmo-community.github.io/ncco-examples/assets/voice_api_audio_streaming.mp3", 5);
+                "{\n" + "  \"message\": \"Stream started\",\n"
+                        + "  \"uuid\": \"ssf61863-4a51-ef6b-11e1-w6edebcf93bb\"\n" + "}"
+        ));
+        StreamResponse response = client.startStream(
+                "ssf61863-4a51-ef6b-11e1-w6edebcf93bb",
+                "https://nexmo-community.github.io/ncco-examples/assets/voice_api_audio_streaming.mp3",
+                5
+        );
         assertEquals("Stream started", response.getMessage());
         assertEquals("ssf61863-4a51-ef6b-11e1-w6edebcf93bb", response.getUuid());
     }
 
     @Test
     public void testStopStream() throws Exception {
-        VoiceClient client = new VoiceClient(stubHttpWrapper(200, "{\n" +
-                "  \"message\": \"Stream stopped\",\n" +
-                "  \"uuid\": \"ssf61863-4a51-ef6b-11e1-w6edebcf93bb\"\n" +
-                "}\n"));
+        VoiceClient client = new VoiceClient(stubHttpWrapper(200,
+                "{\n" + "  \"message\": \"Stream stopped\",\n"
+                        + "  \"uuid\": \"ssf61863-4a51-ef6b-11e1-w6edebcf93bb\"\n" + "}\n"
+        ));
 
         StreamResponse response = client.stopStream("ssf61863-4a51-ef6b-11e1-w6edebcf93bb");
         assertEquals("Stream stopped", response.getMessage());
@@ -236,12 +205,15 @@ public class VoiceClientTest {
     @Test
     public void testStartTalkAllParams() throws Exception {
         VoiceClient client = new VoiceClient(stubHttpWrapper(200,
-                "{\n" +
-                        "  \"message\": \"Talk started\",\n" +
-                        "  \"uuid\": \"ssf61863-4a51-ef6b-11e1-w6edebcf93bb\"\n" +
-                        "}\n"));
+                "{\n" + "  \"message\": \"Talk started\",\n" + "  \"uuid\": \"ssf61863-4a51-ef6b-11e1-w6edebcf93bb\"\n"
+                        + "}\n"
+        ));
 
-        TalkResponse response = client.startTalk("ssf61863-4a51-ef6b-11e1-w6edebcf93bb", "Hello World", VoiceName.CELINE, 8);
+        TalkResponse response = client.startTalk("ssf61863-4a51-ef6b-11e1-w6edebcf93bb",
+                "Hello World",
+                VoiceName.CELINE,
+                8
+        );
         assertEquals("Talk started", response.getMessage());
         assertEquals("ssf61863-4a51-ef6b-11e1-w6edebcf93bb", response.getUuid());
     }
@@ -249,10 +221,9 @@ public class VoiceClientTest {
     @Test
     public void testStartTalkNonLooping() throws Exception {
         VoiceClient client = new VoiceClient(stubHttpWrapper(200,
-                "{\n" +
-                        "  \"message\": \"Talk started\",\n" +
-                        "  \"uuid\": \"ssf61863-4a51-ef6b-11e1-w6edebcf93bb\"\n" +
-                        "}\n"));
+                "{\n" + "  \"message\": \"Talk started\",\n" + "  \"uuid\": \"ssf61863-4a51-ef6b-11e1-w6edebcf93bb\"\n"
+                        + "}\n"
+        ));
 
         TalkResponse response = client.startTalk("ssf61863-4a51-ef6b-11e1-w6edebcf93bb", "Hello World", VoiceName.EMMA);
         assertEquals("Talk started", response.getMessage());
@@ -262,10 +233,9 @@ public class VoiceClientTest {
     @Test
     public void testStartTalkLoopingWithDefaultVoice() throws Exception {
         VoiceClient client = new VoiceClient(stubHttpWrapper(200,
-                "{\n" +
-                        "  \"message\": \"Talk started\",\n" +
-                        "  \"uuid\": \"ssf61863-4a51-ef6b-11e1-w6edebcf93bb\"\n" +
-                        "}\n"));
+                "{\n" + "  \"message\": \"Talk started\",\n" + "  \"uuid\": \"ssf61863-4a51-ef6b-11e1-w6edebcf93bb\"\n"
+                        + "}\n"
+        ));
 
         TalkResponse response = client.startTalk("ssf61863-4a51-ef6b-11e1-w6edebcf93bb", "Hello World", 3);
         assertEquals("Talk started", response.getMessage());
@@ -275,10 +245,9 @@ public class VoiceClientTest {
     @Test
     public void testStartTalkNonLoopingWithDefaultVoice() throws Exception {
         VoiceClient client = new VoiceClient(stubHttpWrapper(200,
-                "{\n" +
-                        "  \"message\": \"Talk started\",\n" +
-                        "  \"uuid\": \"ssf61863-4a51-ef6b-11e1-w6edebcf93bb\"\n" +
-                        "}\n"));
+                "{\n" + "  \"message\": \"Talk started\",\n" + "  \"uuid\": \"ssf61863-4a51-ef6b-11e1-w6edebcf93bb\"\n"
+                        + "}\n"
+        ));
 
         TalkResponse response = client.startTalk("ssf61863-4a51-ef6b-11e1-w6edebcf93bb", "Hello World");
         assertEquals("Talk started", response.getMessage());
@@ -288,10 +257,9 @@ public class VoiceClientTest {
     @Test
     public void testStopTalk() throws Exception {
         VoiceClient client = new VoiceClient(stubHttpWrapper(200,
-                "{\n" +
-                        "  \"message\": \"Talk stopped\",\n" +
-                        "  \"uuid\": \"ssf61863-4a51-ef6b-11e1-w6edebcf93bb\"\n" +
-                        "}\n"));
+                "{\n" + "  \"message\": \"Talk stopped\",\n" + "  \"uuid\": \"ssf61863-4a51-ef6b-11e1-w6edebcf93bb\"\n"
+                        + "}\n"
+        ));
 
         TalkResponse response = client.stopTalk("ssf61863-4a51-ef6b-11e1-w6edebcf93bb");
         assertEquals("Talk stopped", response.getMessage());